+++ test/src/string.test.js
@@ -5,6 +5,7 @@ import isStringBetween  from '../../src/string/isBetween';
 import isNotEmptyString from '../../src/string/isNotEmpty';
 import shorten          from '../../src/string/shorten';
 import humanizeBytes    from '../../src/string/humanizeBytes';
+import humanizeNumber   from '../../src/string/humanizeNumber';
 import {
     fnNumericValues,
     fnBooleanValues,
@@ -27,6 +28,7 @@ const should = chai.should();
 const spy = chai.spy;
 
 describe("String", () => {
+
     describe("isString", () => {
         it ('see a string as a string', () => {
             let vals = fnStringValues();
@@ -445,13 +447,13 @@ describe("String", () => {
             [47328748923747923479, '41.05 EB'],
         ];
 
-        it ('[humanizeBytes] Should return 0 bytes when called with non-alpha-numerical value or 0', () => {
+        it ('Should return 0 bytes when called with non-alpha-numerical value or 0', () => {
             for (const el of [{a:1}, [0,1,2], true, new Date(), /1/g, false, 'hello', 'abc', 0, '0', '-0']) {
                 expect(humanizeBytes(el)).to.eql('0 bytes');
             }
         });
 
-        it ('[humanizeBytes] Should return a positive number between 1 and 1024 (not including 1024) as bytes', () => {
+        it ('Should return a positive number between 1 and 1024 (not including 1024) as bytes', () => {
             for (let i = 1; i < 1024; i++) {
                 if (i < 1000) {
                     expect(humanizeBytes(i)).to.eql(`${i} bytes`);
@@ -463,7 +465,7 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should return a negative number between 1 and 1024 (not including 1024) as bytes', () => {
+        it ('Should return a negative number between 1 and 1024 (not including 1024) as bytes', () => {
             for (let i = -1; i > -1024; i--) {
                 if (i > -1000) {
                     expect(humanizeBytes(i)).to.eql(`${i} bytes`);
@@ -475,7 +477,7 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should return a positive number between 1 and 1024 formatted as string (not including 1024) as bytes', () => {
+        it ('Should return a positive number between 1 and 1024 formatted as string (not including 1024) as bytes', () => {
             for (let i = 1; i < 1024; i++) {
                 if (i < 1000) {
                     expect(humanizeBytes(`${i}`)).to.eql(`${i} bytes`);
@@ -487,7 +489,7 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should return a negative number between 1 and 1024 formatted as string (not including 1024) as bytes', () => {
+        it ('Should return a negative number between 1 and 1024 formatted as string (not including 1024) as bytes', () => {
             for (let i = -1; i > -1024; i--) {
                 if (i > -1000) {
                     expect(humanizeBytes(`${i}`)).to.eql(`${i} bytes`);
@@ -499,23 +501,23 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should correctly convert a positive number', () => {
+        it ('Should correctly convert a positive number', () => {
             for (const el of val_tests) expect(humanizeBytes(el[0])).to.eql(el[1]);
         });
 
-        it ('[humanizeBytes] Should correctly convert a negative number', () => {
+        it ('Should correctly convert a negative number', () => {
             for (const el of val_tests) expect(humanizeBytes(-el[0])).to.eql(`-${el[1]}`);
         });
 
-        it ('[humanizeBytes] Should correctly convert a positive number formatted as string', () => {
+        it ('Should correctly convert a positive number formatted as string', () => {
             for (const el of val_tests) expect(humanizeBytes(`${el[0]}`)).to.eql(el[1]);
         });
 
-        it ('[humanizeBytes] Should correctly convert a negative number formatted as string', () => {
+        it ('Should correctly convert a negative number formatted as string', () => {
             for (const el of val_tests) expect(humanizeBytes(`-${el[0]}`)).to.eql(`-${el[1]}`);
         });
 
-        it ('[humanizeBytes] Should allow overriding precision', () => {
+        it ('Should allow overriding precision', () => {
             const val_tests = [
                 [1024, 10, '1 KB'],
                 [1500, 1, '1.5 KB'],
@@ -543,7 +545,7 @@ describe("String", () => {
 
         });
 
-        it ('[humanizeBytes] Should allow overriding separator', () => {
+        it ('Should allow overriding separator', () => {
             const val_tests = [
                 [1024, 10, '1 KB'],
                 [1500, 1, '1,5 KB'],
@@ -570,7 +572,7 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should allow overriding separator and not trim it', () => {
+        it ('Should allow overriding separator and not trim it', () => {
             const val_tests = [
                 [1024, 10, '1 KB'],
                 [1500, 1, '1 | 5 KB'],
@@ -597,7 +599,7 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should not allow turning off separator', () => {
+        it ('Should not allow turning off separator', () => {
             const val_tests = [
                 [1024, 10, '1 KB'],
                 [1500, 1, '1.5 KB'],
@@ -624,7 +626,7 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should allow overriding delimiter', () => {
+        it ('Should allow overriding delimiter', () => {
             const val_tests = [
                 [1024, 10, '1 KB'],
                 [43244332, 4, '42.230,793 KB'],
@@ -641,7 +643,7 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should allow turning off delimiter by passing as empty string', () => {
+        it ('Should allow turning off delimiter by passing as empty string', () => {
             const val_tests = [
                 [1024, 10, '1 KB'],
                 [43244332, 4, '42230,793 KB'],
@@ -658,7 +660,7 @@ describe("String", () => {
             }
         });
 
-        it ('[humanizeBytes] Should allow overriding units', () => {
+        it ('Should allow overriding units', () => {
             const val_tests = [
                 [20, 10, '20Jedi'],
                 [1024, 10, '1Darth'],
@@ -670,6 +672,263 @@ describe("String", () => {
             for (const el of val_tests) {
                 expect(humanizeBytes(`${el[0]}`, {precision: el[1], units: ['Jedi', 'Darth', 'Vader', 'Force']})).to.eql(el[2]);
                 expect(humanizeBytes(el[0], {precision: el[1], units: ['Jedi', 'Darth', 'Vader', 'Force']})).to.eql(el[2]);
+            }
+        });
+
+    });
+
+    describe("humanizeNumber", () => {
+
+        const val_tests = [
+            [1000, '1k'],
+            [1024, '1.02k'],
+            [1500, '1.5k'],
+            [3584, '3.58k'],
+            [9799, '9.8k'],
+            [58432, '58.43k'],
+            [97432, '97.43k'],
+            [432443, '432.44k'],
+            [857534, '857.53k'],
+            [1000000, '1m'],
+            [1048575, '1.05m'],
+            [5242880, '5.24m'],
+            [1504230, '1.5m'],
+            [3584432, '3.58m'],
+            [9799432, '9.8m'],
+            [584324, '584.32k'],
+            [9743432, '9.74m'],
+            [43244332, '43.24m'],
+            [85753443, '85.75m'],
+            [100000032, '100m'],
+            [1073741823, '1.07b'],
+            [374237489237, '374.24b'],
+            [4893290423489, '4.89t'],
+            [4327963279469432, '4.33q'],
+            [84903298490, '84.9b'],
+            [4903278490, '4.9b'],
+            [438274237890, '438.27b'],
+            [4328904892322, '4.33t'],
+            [974238788, '974.24m'],
+            [47328748923747923479, '47,328.75q'],
+        ];
+
+        it ('Should return 0 when called with non-alpha-numerical value or 0', () => {
+            for (const el of [{a:1}, [0,1,2], true, new Date(), /1/g, false, 'hello', 'abc', 0, '0', '-0']) {
+                expect(humanizeNumber(el)).to.eql('0');
+            }
+        });
+
+        it ('Should return a positive number between 1 and 1000 (not including 1000) without any unit', () => {
+            for (let i = 1; i < 1000; i++) {
+                expect(humanizeNumber(i)).to.eql(`${i}`);
+            }
+        });
+
+        it ('Should return a negative number between 1 and 1000 (not including 1000) without any unit', () => {
+            for (let i = -1; i > -1000; i--) {
+                expect(humanizeNumber(i)).to.eql(`${i}`);
+            }
+        });
+
+        it ('Should return a positive number between 1 and 1000 formatted as string (not including 1000) without any unit', () => {
+            for (let i = 1; i < 1000; i++) {
+                expect(humanizeNumber(`${i}`)).to.eql(`${i}`);
+            }
+        });
+
+        it ('Should return a negative number between 1 and 1000 formatted as string (not including 1000) without any unit', () => {
+            for (let i = -1; i > -1000; i--) {
+                expect(humanizeNumber(`${i}`)).to.eql(`${i}`);
+            }
+        });
+
+        it ('Should correctly convert a positive number', () => {
+            for (const el of val_tests) expect(humanizeNumber(el[0])).to.eql(el[1]);
+        });
+
+        it ('Should correctly convert a negative number', () => {
+            for (const el of val_tests) expect(humanizeNumber(-el[0])).to.eql(`-${el[1]}`);
+        });
+
+        it ('Should correctly convert a positive number formatted as string', () => {
+            for (const el of val_tests) expect(humanizeNumber(`${el[0]}`)).to.eql(el[1]);
+        });
+
+        it ('Should correctly convert a negative number formatted as string', () => {
+            for (const el of val_tests) expect(humanizeNumber(`-${el[0]}`)).to.eql(`-${el[1]}`);
+        });
+
+        it ('Should allow overriding precision', () => {
+            const val_tests = [
+                [1024, 10, '1.024k'],
+                [1500, 1, '1.5k'],
+                [1500, 2, '1.5k'],
+                [1500, 3, '1.5k'],
+                [3584, 1, '3.6k'],
+                [3584, 2, '3.58k'],
+                [3584, 3, '3.584k'],
+                [9799, 1, '9.8k'],
+                [9799, 4, '9.799k'],
+                [58432, 1, '58.4k'],
+                [58432, 2, '58.43k'],
+                [58432, 3, '58.432k'],
+                [4893290423489, 5, '4.89329t'],
+                [4893290423489, 4, '4.8933t'],
+                [4893290423489, 3, '4.893t'],
+                [4893290423489, 2, '4.89t'],
+                [4893290423489, 1, '4.9t'],
+                [4893290423489, 0, '5t'],
+            ];
+            for (const el of val_tests) {
+                expect(humanizeNumber(`${el[0]}`, {precision: el[1]})).to.eql(el[2]);
+                expect(humanizeNumber(el[0], {precision: el[1]})).to.eql(el[2]);
+            }
+
+        });
+
+        it ('Should allow overriding separator', () => {
+            const val_tests = [
+                [1024, 10, '1,024k'],
+                [1500, 1, '1,5k'],
+                [1500, 2, '1,5k'],
+                [1500, 3, '1,5k'],
+                [3584, 1, '3,6k'],
+                [3584, 2, '3,58k'],
+                [3584, 3, '3,584k'],
+                [9799, 1, '9,8k'],
+                [9799, 4, '9,799k'],
+                [58432, 1, '58,4k'],
+                [58432, 2, '58,43k'],
+                [58432, 3, '58,432k'],
+                [4893290423489, 5, '4,89329t'],
+                [4893290423489, 4, '4,8933t'],
+                [4893290423489, 3, '4,893t'],
+                [4893290423489, 2, '4,89t'],
+                [4893290423489, 1, '4,9t'],
+                [4893290423489, 0, '5t'],
+            ];
+            for (const el of val_tests) {
+                expect(humanizeNumber(`${el[0]}`, {precision: el[1], separator: ','})).to.eql(el[2]);
+                expect(humanizeNumber(el[0], {precision: el[1], separator: ','})).to.eql(el[2]);
+            }
+        });
+
+        it ('Should allow overriding separator and not trim it', () => {
+            const val_tests = [
+                [1024, 10, '1 | 024k'],
+                [1500, 1, '1 | 5k'],
+                [1500, 2, '1 | 5k'],
+                [1500, 3, '1 | 5k'],
+                [3584, 1, '3 | 6k'],
+                [3584, 2, '3 | 58k'],
+                [3584, 3, '3 | 584k'],
+                [9799, 1, '9 | 8k'],
+                [9799, 4, '9 | 799k'],
+                [58432, 1, '58 | 4k'],
+                [58432, 2, '58 | 43k'],
+                [58432, 3, '58 | 432k'],
+                [4893290423489, 5, '4 | 89329t'],
+                [4893290423489, 4, '4 | 8933t'],
+                [4893290423489, 3, '4 | 893t'],
+                [4893290423489, 2, '4 | 89t'],
+                [4893290423489, 1, '4 | 9t'],
+                [4893290423489, 0, '5t'],
+            ];
+            for (const el of val_tests) {
+                expect(humanizeNumber(`${el[0]}`, {precision: el[1], separator: ' | '})).to.eql(el[2]);
+                expect(humanizeNumber(el[0], {precision: el[1], separator: ' | '})).to.eql(el[2]);
+            }
+        });
+
+        it ('Should not allow turning off separator', () => {
+            const val_tests = [
+                [1024, 10, '1.024k'],
+                [1500, 1, '1.5k'],
+                [1500, 2, '1.5k'],
+                [1500, 3, '1.5k'],
+                [3584, 1, '3.6k'],
+                [3584, 2, '3.58k'],
+                [3584, 3, '3.584k'],
+                [9799, 1, '9.8k'],
+                [9799, 4, '9.799k'],
+                [58432, 1, '58.4k'],
+                [58432, 2, '58.43k'],
+                [58432, 3, '58.432k'],
+                [4893290423489, 5, '4.89329t'],
+                [4893290423489, 4, '4.8933t'],
+                [4893290423489, 3, '4.893t'],
+                [4893290423489, 2, '4.89t'],
+                [4893290423489, 1, '4.9t'],
+                [4893290423489, 0, '5t'],
+            ];
+            for (const el of val_tests) {
+                expect(humanizeNumber(`${el[0]}`, {precision: el[1], separator: ''})).to.eql(el[2]);
+                expect(humanizeNumber(el[0], {precision: el[1], separator: ''})).to.eql(el[2]);
+            }
+        });
+
+        it ('Should allow overriding delimiter', () => {
+            const val_tests = [
+                [1024, 10, '1,024K'],
+                [43244332, 4, '43.244,332K'],
+                [85753443, 3, '85.753,443K'],
+                [100000032, 5, '100.000,032K'],
+                [1073741823, 2, '1.073.741,82K'],
+                [374237489237, 1, '374.237.489,2K'],
+                [4893290423489, 0, '4.893.290.423K'],
+                [4327963279469432, 5, '4.327.963.279.469,432K'],
+            ];
+            for (const el of val_tests) {
+                expect(humanizeNumber(`${el[0]}`, {precision: el[1], units: ['', 'K'], delim: '.', separator: ','})).to.eql(el[2]);
+                expect(humanizeNumber(el[0], {precision: el[1], units: ['', 'K'], delim: '.', separator: ','})).to.eql(el[2]);
+            }
+        });
+
+        it ('Should allow turning off delimiter by passing as empty string', () => {
+            const val_tests = [
+                [1024, 10, '1,024K'],
+                [43244332, 4, '43244,332K'],
+                [85753443, 3, '85753,443K'],
+                [100000032, 5, '100000,032K'],
+                [1073741823, 2, '1073741,82K'],
+                [374237489237, 1, '374237489,2K'],
+                [4893290423489, 0, '4893290423K'],
+                [4327963279469432, 5, '4327963279469,432K'],
+            ];
+            for (const el of val_tests) {
+                expect(humanizeNumber(`${el[0]}`, {precision: el[1], units: ['', 'K'], delim: '', separator: ','})).to.eql(el[2]);
+                expect(humanizeNumber(el[0], {precision: el[1], units: ['', 'K'], delim: '', separator: ','})).to.eql(el[2]);
+            }
+        });
+
+        it ('Should allow overriding units', () => {
+            const val_tests = [
+                [20, 10, '20Jedi'],
+                [1024, 10, '1.024Darth'],
+                [43244332, 4, '43.2443Vader'],
+                [85753443, 3, '85.753Vader'],
+                [4893290423489, 0, '4,893Force'],
+                [4327963279469432, 5, '4,327,963.27947Force'],
+            ];
+            for (const el of val_tests) {
+                expect(humanizeNumber(`${el[0]}`, {precision: el[1], units: ['Jedi', 'Darth', 'Vader', 'Force']})).to.eql(el[2]);
+                expect(humanizeNumber(el[0], {precision: el[1], units: ['Jedi', 'Darth', 'Vader', 'Force']})).to.eql(el[2]);
+            }
+        });
+
+        it ('Should allow simply humanizing without units', () => {
+            const val_tests = [
+                [0, 10, '0'],
+                [20, 10, '20'],
+                [1024, 10, '1,024'],
+                [43244332.43244324, 4, '43,244,332.4324'],
+                [85753443, 3, '85,753,443'],
+                [4893290423489, 0, '4,893,290,423,489'],
+                [3279469432.424236, 5, '3,279,469,432.42424'],
+            ];
+            for (const el of val_tests) {
+                expect(humanizeNumber(`${el[0]}`, {precision: el[1], units: false})).to.eql(el[2]);
+                expect(humanizeNumber(el[0], {precision: el[1], units: false})).to.eql(el[2]);
             }
         });