+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 ### Added
+- string/humanizeNumber (with delim, separator, precision, units, divider) options
 - number/isAboveOrEqual
 - number/isBelowOrEqual
 - number/isIntegerAboveOrEqual